requirements.txt
# API客户端（可选，使用DeepSeek时不需要）
anthropic>=0.18.0

# arXiv官方库
arxiv>=2.0.0

# HTTP请求
requests>=2.28.0
aiohttp>=3.8.0

# HTML解析
beautifulsoup4>=4.11.0

# 配置管理
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# 数据处理
pyyaml>=6.0

# 重试机制
tenacity>=8.0.0

# 日志
loguru>=0.6.0
